Move loading and combine code into their own submodules.
